+++ pkg/k8s/auth.go
@@ -1,9 +1,13 @@
 package k8s
 
 import (
+	"context"
 	"fmt"
 
+	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
+	authv1 "k8s.io/api/authorization/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 const (
@@ -18,6 +22,24 @@ const (
   * Update your kubeconfig by following https://docs.aws.amazon.com/eks/latest/userguide/create-kubeconfig.html
 `
 )
+
+func (kubeClient *Client) isActionPermitted(ctx context.Context, action *authv1.ResourceAttributes) (bool, error) {
+	var err error
+
+	accessReview := &authv1.SelfSubjectAccessReview{
+		Spec: authv1.SelfSubjectAccessReviewSpec{ResourceAttributes: action},
+	}
+
+	if accessReview, err = kubeClient.AuthorizationV1().SelfSubjectAccessReviews().Create(ctx, accessReview, metav1.CreateOptions{}); err != nil {
+		return false, errors.Wrapf(err, "api error on resource: %s", action.Resource)
+	}
+
+	if accessReview.Status.Denied || !accessReview.Status.Allowed {
+		return false, nil
+	}
+
+	return true, nil
+}
 
 func (kubeClient *Client) printHintIfAuthError(err error) error {
 	switch err.Error() {